+++ hack/.golint_failures
@@ -111,7 +111,6 @@ pkg/kubectl/cmd/convert
 pkg/kubectl/cmd/get
 pkg/kubelet/apis/config
 pkg/kubelet/apis/config/v1beta1
-pkg/kubelet/apis/deviceplugin/v1beta1
 pkg/kubelet/checkpointmanager/testing/example_checkpoint_formats/v1
 pkg/kubelet/cm
 pkg/kubelet/config
