+++ pkg/kubelet/pluginmanager/operationexecutor/operation_generator.go
@@ -31,7 +31,7 @@ import (
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"k8s.io/client-go/tools/record"
-	registerapi "k8s.io/kubernetes/pkg/kubelet/apis/pluginregistration/v1"
+	registerapi "k8s.io/kubelet/pkg/apis/pluginregistration/v1"
 	"k8s.io/kubernetes/pkg/kubelet/pluginmanager/cache"
 )
