+++ staging/src/k8s.io/kubelet/pkg/apis/pluginregistration/v1/BUILD
@@ -11,13 +11,12 @@ go_library(
         "api.pb.go",
         "constants.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/kubelet/apis/pluginregistration/v1alpha1",
+    importpath = "k8s.io/kubelet/pkg/apis/pluginregistration/v1",
     deps = [
         "//vendor/github.com/gogo/protobuf/gogoproto:go_default_library",
         "//vendor/github.com/gogo/protobuf/proto:go_default_library",
+        "//vendor/golang.org/x/net/context:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
-        "//vendor/google.golang.org/grpc/codes:go_default_library",
-        "//vendor/google.golang.org/grpc/status:go_default_library",
     ],
 )
