+++ test/images/sample-device-plugin/sampledeviceplugin.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"k8s.io/klog"
-	pluginapi "k8s.io/kubernetes/pkg/kubelet/apis/deviceplugin/v1beta1"
+	pluginapi "k8s.io/kubelet/pkg/apis/deviceplugin/v1beta1"
 	dm "k8s.io/kubernetes/pkg/kubelet/cm/devicemanager"
 )
